pyproject.toml
[project]
name = "howtoinvestwithai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cozepy>=0.13.0",
    "jupyter-book>=2.0.0a0",
    "python-dotenv>=1.0.1",
]

[dependency-groups]
dev = [
    "ruff>=0.11.0",
]
